[rfc2217] Wrap unsupport ed methods of rfc2217 implementation
The serial io implementation of openHAB hides the implementation details to the module using serial io.
However the telnet implementation throws on some method an UnsupportedException when called.
This results in a leak of abstraction in that when using such method the implementation must make a difference between telnet and other connections.
Therefore this change wraps the specific methods and simply does nothing.
This fixes eclipse-archived/smarthome/issues/6280

Signed-off-by: Hilbrand Bouwkamp <hilbrand@h72.nl>